Rearrange state package a bit, add some code to look up the right state resolution algorithm
